Fixed SonarQube comments
