+++ server/tests/internal/test_internal_api_secret.py
@@ -18,7 +18,7 @@ _server_dir = os.path.join(os.path.dirname(__file__), os.pardir, os.pardir)
 if os.path.abspath(_server_dir) not in sys.path:
     sys.path.insert(0, os.path.abspath(_server_dir))
 
-import utils.internal.api_handler
+from utils.internal import api_handler as _api_handler_module
 from utils.internal.api_handler import _handle_kubectl_execute  # noqa: E402
 
 
@@ -173,7 +173,7 @@ class TestConstantTimeComparison:
         so the runtime does not leak the secret length or prefix via timing.
         """
         compare_spy = MagicMock(wraps=hmac.compare_digest)
-        monkeypatch.setattr(utils.internal.api_handler.hmac, "compare_digest", compare_spy)
+        monkeypatch.setattr(_api_handler_module.hmac, "compare_digest", compare_spy)
 
         writer = _writer_spy()
         asyncio.run(
